Moved code for saving AVI and WAV format files into new file_save.c
