+++ src/Makefile.am
@@ -123,6 +123,9 @@ endif
 if WANT_POKEYREC
 atari800_SOURCES += pokeyrec.c pokeyrec.h
 endif
+if WITH_FILE_SAVE
+atari800_SOURCES += file_save.c file_save.h
+endif
