+++ src/file_save.h
@@ -1,0 +1,24 @@
+#ifndef FILE_SAVE_H_
+#define FILE_SAVE_H_
+
+#include "atari.h"
+
+void fputw(UWORD, FILE *fp);
+void fputl(ULONG, FILE *fp);
+size_t fwritele(const void *ptr, size_t size, size_t nmemb, FILE *fp);
+
+#ifdef SOUND
+FILE *WAV_OpenFile(const char *szFileName);
+int WAV_WriteSamples(const unsigned char *buf, unsigned int num_samples, FILE *fp);
+int WAV_CloseFile(FILE *fp);
+#ifdef AVI_VIDEO_RECORDING
+int MRLE_CreateFrame(UBYTE *buf, int bufsize, const UBYTE *source);
+FILE *AVI_OpenFile(const char *szFileName);
+int AVI_CloseFile(FILE *fp);
+int AVI_AddVideoFrame(FILE *fp);
+int AVI_AddAudioSamples(const UBYTE *buf, int num_samples, FILE *fp);
+#endif /* AVI_VIDEO_RECORDING */
+#endif /* SOUND */
+
+#endif /* FILE_SAVE_H_ */
+
